chore: extract repeated logic to formatTime function
this slightly changes the formatting of the original use (it only pads minutes when time >= 1 hour), so it's not strictly a refactor, but I doubt there's an issue with that.
