+++ components/user/card-seekbar.vue
@@ -11,6 +11,8 @@
 </template>
   
 <script setup lang="ts">
+import { formatTime } from '~/utils/format-time';
+
 const props = defineProps<{
     startTimeMs: number, 
     endTimeMs: number
@@ -21,20 +23,5 @@ const currentDateTime = useNow();
 const currentPosition = computed(() => formatTime(props.startTimeMs, currentDateTime.value.getTime()));
 const songDuration = computed(() => formatTime(props.startTimeMs, props.endTimeMs));
 const percentage = computed(() => ((currentDateTime.value.getTime() - props.startTimeMs) / (props.endTimeMs - props.startTimeMs)) * 100);
-
-const formatTime = (startTimeMs: number, endTimeMs: number) => {
-    const secondAsMilliseconds = 1000;
-    const minuteAsMilliseconds = secondAsMilliseconds * 60;
-    const hourAsMilliseconds = minuteAsMilliseconds * 60;
-
-    const distance = endTimeMs - startTimeMs;
-    const seconds = Math.floor((distance / secondAsMilliseconds) % 60).toString().padStart(2, '0');
-    let minutes = Math.floor((distance / minuteAsMilliseconds) % 60).toString()
-    if (distance < hourAsMilliseconds) return `${minutes}:${seconds}`;
-
-    minutes = minutes.padStart(2, '0');
-    const hours = Math.floor(distance / hourAsMilliseconds).toString()
-    return `${hours}:${minutes}:${seconds}`;
-}
   
 </script>
